+++ srcs/requirements/wordpress/tools/installwp.sh
@@ -1,4 +1,20 @@
 #!/bin/bash
+if [ ! -f /run/secrets/wp_db_password ]; then
+	echo "No secret found for wp_db_password"
+	exit 1
+fi
+if [ ! -f /run/secrets/wp_user_password ]; then
+	echo "No secret found for wp_user_password"
+	exit 1
+fi
+if [ ! -f /run/secrets/wp_admin_password ]; then
+	echo "No secret found for wp_admin_password"
+	exit 1
+fi
+
+WP_DB_PASSWORD = $(cat /run/secrets/wp_db_password)
+WP_ADMIN_PASSWORD = $(cat /run/secrets/wp_admin_password)
+
 cd /var/www/html
 curl -o /usr/local/bin/wp -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
 chmod +x /usr/local/bin/wp
@@ -16,24 +32,27 @@ done
 wp config create \
 	--dbname=$WP_DB_NAME \
 	--dbuser=$WP_DB_USER \
-	--dbpass=$WP_DB_PASSWORD \
 	--dbhost=$WP_DB_HOST \
-	--allow-root
+	--allow-root \
+	--quiet \
+	--prompt=dbpass < /run/secrets/wp_db_password
 
 wp core install \
 	--url=$DOMAIN \
 	--title=$WP_TITLE \
 	--admin_user=$WP_ADMIN_USER \
-	--admin_password=$WP_ADMIN_PASSWORD \
 	--admin_email=$WP_ADMIN_EMAIL \
-	--allow-root
+	--allow-root \
+	--quiet \
+	--prompt=admin_password < /run/secrets/wp_admin_password
 
 wp user create \
 	$WP_USER \
 	$WP_USER_EMAIL \
 	--role=author \
-	--user_pass=$WP_USER_PASSWORD \
-	--allow-root
+	--allow-root \
+	--quiet \
+	--prompt=user_pass < /run/secrets/wp_user_password
 
 wp plugin install \
 	redis-cache \
